+++ classec/service/ec2/Ec2Service.py
@@ -2,11 +2,7 @@ import timeit
 import boto3
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-
 from Ec2Schema import ec2DataWithCosting,Base, ec2Costing
-
-
-#engine = create_engine('mysql://root:root@110.110.110.170/test_govind')
 
 def ToCheckForNextMarker(retrieveNextEC2):
     if 'NextToken' in retrieveNextEC2:
@@ -19,6 +15,8 @@ def ToCheckForNextMarker(retrieveNextEC2):
 def ToLoadDataOntoFile(Datatoload):
     ListOfAllInst = Datatoload['Reservations']  # getting all instances
     print(len(ListOfAllInst))
+    counter = 0
+    recordBuilder = list()
     for Inst in ListOfAllInst:
         EachInst = Inst['Instances']
         for instDetails in EachInst:
@@ -27,7 +25,6 @@ def ToLoadDataOntoFile(Datatoload):
             ec2_type = instDetails['InstanceType']
             ec2_state = instDetails['State']['Name']
             ec2_launchdate = (instDetails['LaunchTime']).date()
-            #ec2_launchdate = (dateutil.parser.parse(instDetails['LaunchTime'])).strftime('%Y-%m-%d')
             ec2_monitoring = instDetails['Monitoring']['State']
             ec2_reserveid = ''
             if 'ReservationId' in Inst:
@@ -58,11 +55,6 @@ def ToLoadDataOntoFile(Datatoload):
                     if tag['Key'] == 'Name':
                         ec2_name = tag['Value']
 
-#            query = "insert into "+tbName+" values ('"+ec_id+"','"+ec2_amiid+"','"+ec2_type+"','"+ec2_state+"','"+ec2_domain+"','"+ec2_env+"','"+ec2_name+"','"+ec2_region+"','"+str(ec2_launchdate)+"','"+ec2_monitoring+"','"+ec2_reserveid+"','"+ec2_subnet+"','"+ec2_hypervisor+"','"+ec2_vpc+"','"+str(ec2_EbsOptimized)+"','"+ec2_Architecture+"','"+ec2_platform+"')"
-#            session.add(ec2DataWithCosting(ec_id, ec2_amiid, ec2_type,ec2_state, ec2_domain,ec2_env, ec2_name,
-#                                            ec2_region, str(ec2_launchdate), ec2_monitoring, ec2_reserveid, ec2_subnet,
-#                                            ec2_hypervisor,ec2_vpc,ec2_EbsOptimized, ec2_Architecture,ec2_platform))
-
             cost = 0
             for eachEc2record in CostingData:
                 if ec2_type == eachEc2record.apiname and ec2_region[:-1] == eachEc2record.Region:
@@ -77,11 +69,21 @@ def ToLoadDataOntoFile(Datatoload):
                         else:
                             cost = eachEc2record.cost_ondemand_linux_hourly
 
-                    session.add(ec2DataWithCosting(ec_id, ec2_amiid, ec2_type, ec2_state, ec2_domain, ec2_env, ec2_name,
-                                                   ec2_region, str(ec2_launchdate), ec2_monitoring, ec2_reserveid,
-                                                   ec2_subnet,ec2_hypervisor, ec2_vpc, ec2_EbsOptimized, ec2_Architecture,
-                                                   ec2_platform,cost))
+            counter = counter + 1
+            record = ec2DataWithCosting(ec_id, ec2_amiid, ec2_type, ec2_state, ec2_domain, ec2_env, ec2_name,
+                                        ec2_region, str(ec2_launchdate), ec2_monitoring, ec2_reserveid,
+                                        ec2_subnet,ec2_hypervisor, ec2_vpc, ec2_EbsOptimized, ec2_Architecture,
+                                        ec2_platform,cost)
+            recordBuilder.append(record)
 
+            if 500 == counter:
+                session.bulk_save_objects(recordBuilder)
+                session.commit()
+                counter = 0
+                recordBuilder = []
+
+    session.bulk_save_objects(recordBuilder)
+    session.commit()
 
 
 def gettingEc2Costing(ModelClassName=None):
@@ -98,8 +100,9 @@ if __name__ == '__main__':
     DBsession = sessionmaker(bind=engine)
     session = DBsession()
 
-    print "Creation table ebs_static_data"
-    ec2DataWithCosting.__table__.drop()
+    print "Creation table ec2_static_data"
+    if engine.dialect.has_table(engine, ec2DataWithCosting.__tablename__):
+        ec2DataWithCosting.__table__.drop()
     ec2DataWithCosting.__table__.create()
 
     print 'getting Data'
@@ -116,6 +119,5 @@ if __name__ == '__main__':
         DescEC2 = EC2.describe_instances()
         ToCheckForNextMarker(DescEC2)
 
-    session.commit()
     elapsed = timeit.default_timer() - start_time
     print elapsed
